requirements.txt
cffi==1.16.0
click==7.1.2  # replace local reference with latest version or a version compatible with your project
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
filelock==3.13.4
Flask==1.1.2  # replace local reference with latest version or a version compatible with your project
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
flwr==1.8.0
fonttools==4.51.0
fsspec==2024.3.1
grpcio==1.62.1
iterators==0.0.2
itsdangerous==1.1.0  # replace local reference with latest version or a version compatible with your project
Jinja2==2.11.3  # replace local reference with latest version or a version compatible with your project
joblib==1.4.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==1.1.1  # replace local reference with latest version or a version compatible with your project
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sympy==1.12
threadpoolctl==3.4.0
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
triton==2.2.0
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
Werkzeug==1.0.1  # replace local reference with latest version or a version compatible with your project
